--- SS14.Client/ResourceManagement/ResourceTypes/JsonResource.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SS14.Client.ResourceManagement.ResourceTypes
-{
-    class JsonResource
-    {
-    }
-}
